requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-generativeai
pypdf
python-multipart
python-jose
passlib[bcrypt]